test: e2e - add memory snapshot checks for other tests
